requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
twilio==9.0.4
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
requests==2.31.0
boto3==1.34.0
django-storages==1.14.2
whitenoise==6.6.0
